tunerhal: fix compile error on 64-bit [1/1]
PD#SWPL-180663

Problem:
tunerhal support 64-bit compile

Solution:
fix compile error on 64-bit

Verify:
ohm

+++ aidl/MultiHwDemux/AmDmx.cpp
@@ -704,7 +704,7 @@ AM_ErrorCode_t AM_DMX_Device::AM_DMX_GetFilterFd(int fhandle, int *fd) {
     ret = dmx_get_used_filter(fhandle, &filter);
 
     if (ret == AM_SUCCESS && fd) {
-        *fd = (int)filter->drv_data;
+        *fd = (uintptr_t)filter->drv_data;
         ret = AM_SUCCESS;
     }
     pthread_mutex_unlock(&lock);
